refactor: move via-related ble code to own module
